fix: moved parts of report module to simulator module for better consistency
