+++ oopnet/report/__init__.py
@@ -1,39 +1,1 @@
-from .simulation_errors import (
-    EPANETSimulationError,
-    EPANETError,
-    InsufficientMemoryError,
-    HydraulicEquationError,
-    InputDataError,
-    EPANETSyntaxError,
-    IllegalNumericalValueError,
-    UndefinedNodeError,
-    UndefinedLinkError,
-    UndefinedTimePatternError,
-    UndefinedCurveError,
-    CheckValveImmutableError,
-    NodeReferenceError,
-    IllegalNodePropertyError,
-    LinkReferenceError,
-    IllegalLinkPropertyError,
-    UndefinedTraceNodeError,
-    IllegalAnalysisOptionError,
-    InvalidIDError,
-    SharedIDError,
-    UndefinedPumpError,
-    InvalidEnergyDataError,
-    IllegalValveValveConnectionError,
-    IllegalValveSourceConnectionError,
-    MisplacedRuleClause,
-    NotEnoughNodesError,
-    NotEnoughSourcesError,
-    InvalidTankLevelError,
-    InvalidPumpError,
-    InvalidCurveError,
-    UnconnectedNodeError,
-    TempInputFileAccessError,
-    ReportFileAccessError,
-    BinaryOutputFileAccessError,
-    ResultFileSavingError,
-    ReportFileSavingError,
-)
 from .report import SimulationReport
