+++ oopnet/simulator/__init__.py
@@ -1,2 +1,38 @@
-# from .epanet2 import run as Run
-# from .epanet2 import ModelSimulator
+from .simulation_errors import (
+    EPANETSimulationError,
+    EPANETError,
+    InsufficientMemoryError,
+    HydraulicEquationError,
+    InputDataError,
+    EPANETSyntaxError,
+    IllegalNumericalValueError,
+    UndefinedNodeError,
+    UndefinedLinkError,
+    UndefinedTimePatternError,
+    UndefinedCurveError,
+    CheckValveImmutableError,
+    NodeReferenceError,
+    IllegalNodePropertyError,
+    LinkReferenceError,
+    IllegalLinkPropertyError,
+    UndefinedTraceNodeError,
+    IllegalAnalysisOptionError,
+    InvalidIDError,
+    SharedIDError,
+    UndefinedPumpError,
+    InvalidEnergyDataError,
+    IllegalValveValveConnectionError,
+    IllegalValveSourceConnectionError,
+    MisplacedRuleClause,
+    NotEnoughNodesError,
+    NotEnoughSourcesError,
+    InvalidTankLevelError,
+    InvalidPumpError,
+    InvalidCurveError,
+    UnconnectedNodeError,
+    TempInputFileAccessError,
+    ReportFileAccessError,
+    BinaryOutputFileAccessError,
+    ResultFileSavingError,
+    ReportFileSavingError,
+)
