+++ oopnet/simulator/binaryfile_reader.py
@@ -1,5 +1,4 @@
 import logging
-
 import numpy as np
 
 logger = logging.getLogger(__name__)
