+++ oopnet/simulator/reportfile_reader.py
@@ -7,7 +7,7 @@ import pandas as pd
 import xarray as xr
 from xarray import DataArray, Dataset
 
-from oopnet.report.error_manager import ErrorManager
+from oopnet.simulator.error_manager import ErrorManager
 from oopnet.utils.oopnet_logging import logging_decorator
 
 logger = logging.getLogger(__name__)
